requirements.txt
GitPython
Faker
markovify